Set mark
Set a mark, this mark could be used to refactor
